# gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  public.GetAllPartyResponse:
    properties:
      parties:
        items:
          $ref: '#/definitions/public.Party'
        type: array
    type: object
...
  public.Party:
    properties:
      created_at:
        type: string
      deleted_at:
        type: integer
      description:
        type: string
      id:
        type: string
      name:
        type: string
      opened_date:
        type: string
      slogan:
        type: string
      updated_at:
        type: string
    type: object
  public.PartyCreate:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      opened_date:
        type: string
      slogan:
        type: string
    type: object
  public.PartyUpdate:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      opened_date:
        type: string
      slogan:
        type: string
    type: object
...
  vote.Candidate:
    properties:
      created_at:
        type: string
      deleted_at:
        type: integer
      election_id:
        type: string
      id:
        type: string
      public_id:
        type: string
      updated_at:
        type: string
    type: object
  vote.CandidateCreate:
    properties:
      election_id:
        type: string
      public_id:
        type: string
    type: object
  vote.CandidateDelete:
    properties:
      id:
        type: string
    type: object
...
  vote.CandidateUpdate:
    properties:
      election_id:
        type: string
      id:
        type: string
      public_id:
        type: string
    type: object
...
  vote.GetAllCandidateRes:
    properties:
      candidates:
        items:
          $ref: '#/definitions/vote.Candidate'
        type: array
      count:
        type: integer
    type: object
...
  vote.Void:
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing online voting system resources
  title: Online Voting System API
  version: "1.0"
paths:
  /candidate:
    delete:
      consumes:
      - application/json
      description: Deletes an existing candidate by ID.
      parameters:
      - description: Candidate data
        in: body
        name: candidate
        required: true
        schema:
          $ref: '#/definitions/vote.CandidateDelete'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.Void'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete an existing candidate
      tags:
      - Candidates
    post:
      consumes:
      - application/json
      description: Creates a new candidate for an election.
      parameters:
      - description: Candidate data
        in: body
        name: candidate
        required: true
        schema:
          $ref: '#/definitions/vote.CandidateCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.Candidate'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new candidate
      tags:
      - Candidates
    put:
      consumes:
      - application/json
      description: Updates an existing candidate.
      parameters:
      - description: Candidate data
        in: body
        name: candidate
        required: true
        schema:
          $ref: '#/definitions/vote.CandidateUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.Void'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update an existing candidate
      tags:
      - Candidates
  /candidate/all:
    get:
      consumes:
      - application/json
      description: Retrieves all candidates associated with an election.
      parameters:
      - description: Election ID
        in: query
        name: election_id
        type: string
      - description: Public ID (optional)
        in: query
        name: public_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.GetAllCandidateRes'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all candidates
      tags:
      - Candidates
  /candidate/id:
    get:
      consumes:
      - application/json
      description: Retrieves a candidate by its ID.
      parameters:
      - description: Candidate ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.Candidate'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a candidate by its ID
      tags:
      - Candidates
...
  /party:
    delete:
      consumes:
      - application/json
      description: Deletes a party by its ID.
      parameters:
      - description: Party ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.Void'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a party
      tags:
      - Parties
    post:
      consumes:
      - application/json
      description: Creates a new party.
      parameters:
      - description: Party data
        in: body
        name: party
        required: true
        schema:
          $ref: '#/definitions/public.PartyCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/public.Party'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new party
      tags:
      - Parties
    put:
      consumes:
      - application/json
      description: Updates an existing party.
      parameters:
      - description: Party data
        in: body
        name: party
        required: true
        schema:
          $ref: '#/definitions/public.PartyUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vote.Void'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update a party
      tags:
      - Parties
  /party/all:
    get:
      consumes:
      - application/json
      description: Retrieves all parties.
      parameters:
      - description: Opened date (optional)
        in: query
        name: opened_date
        type: string
      - description: Name (optional)
        in: query
        name: name
        type: string
      - description: Slogan (optional)
        in: query
        name: slogan
        type: string
      - description: Description (optional)
        in: query
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/public.GetAllPartyResponse'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all parties
      tags:
      - Parties
  /party/id:
    get:
      consumes:
      - application/json
      description: Retrieves a party by its ID.
      parameters:
      - description: Party ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/public.Party'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a party by its ID
      tags:
      - Parties
...
swagger: "2.0"